refactor:#190 supabase 타입 상수 적용
