+++ src/components/features/home/main-gonggam-item.tsx
@@ -1,7 +1,7 @@
 import ItemContainer from '@/components/features/home/main-gonggam-item-container';
 import ContentsContainer from '@/components/features/home/main-gonggam-item-contents-container';
 import ReactionBox from '@/components/features/home/main-gonggam-item-reaction-box';
-import type { GonggamPostWithReaction } from '@/types/gonggam-posts';
+import type { GonggamPostWithReaction } from '@/types/gonggam';
 
 interface ItemProps {
   post: GonggamPostWithReaction;
