+++ src/components/features/mypage/button-client-component.tsx
@@ -3,11 +3,11 @@
 import Image from 'next/image';
 import { useState } from 'react';
 import RandomMissionModal from '@/components/features/modals/random-mission';
-import type { Tables } from '@/types/supabase';
 import CHEVRON_RIGHT from '@images/images/chevron-right.svg';
+import type { TableMissionList } from '@/types/supabase-const';
 
 interface ButtonClientComponentProps {
-  missionsData: Tables<'mission_list'>[];
+  missionsData: TableMissionList[];
   isLogin: boolean;
 }
