+++ src/components/layout/header/header-nav-challenge-menu.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import HeaderDropdownMenu from '@/components/layout/header/header-dropdown-menu';
 import { PATH } from '@/constants/page-path';
-import type { MenuItem } from '@/types/components/header';
+import type { MenuItem } from '@/types/header';
 import MENU_ICON from '@images/images/challenge_menu.svg';
 
 const NavChallengeMenu = () => {
