+++ src/lib/hooks/queries/use-get-gonggam-posts-infinite-query.ts
@@ -1,7 +1,7 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { QUERY_KEY } from '@/constants/query-keys';
 import { getMyGonggamPostsAll } from '@/lib/utils/api/my-gonggam-client.api';
-import type { GetMyGonggamPostsResponse, SortBy } from '@/types/gonggam-posts';
+import type { GetMyGonggamPostsResponse, SortBy } from '@/types/gonggam';
 
 /**
  * @function useGetGonggamPostsInfiniteQuery
