+++ src/lib/hooks/queries/use-get-user-profile.ts
@@ -1,9 +1,9 @@
 import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEY } from '@/constants/query-keys';
 import { fetchUserProfile } from '@/lib/utils/api/auth-client.api';
-import type { Tables } from '@/types/supabase';
+import type { TableUsers } from '@/types/supabase-const';
 
-export const useUserProfile = (initProfile: Tables<'users'>) => {
+export const useUserProfile = (initProfile: TableUsers) => {
   const {
     data: profile,
     isPending: isProfilePending,
