+++ src/lib/utils/api/auth-action.ts
@@ -66,7 +66,7 @@ export const logout = async () => {
  * 사용자가 프로필 정보를 업데이트 하는 api 함수입니다.
  * 현재 로그인 된 사용자 아이디를 받아와서 아이디에 해당하는 row를 파라미터로 받은 정보로 변경합니다.
  * @param { { string, string } } - 변경할 닉네임과 프로필 이미지기 업로드 된 주소
- * @returns { Tables<'users'> } - 현재 세션에 해당하는 users 테이블 row
+ * @returns { TableUser } - 현재 세션에 해당하는 users 테이블 row
  */
 export const updateUserProfile = async (formData: { nickname: string; profile_image: string }): Promise<void> => {
   const supabase = await createClient();
