+++ src/lib/utils/api/my-gonggam-client.api.ts
@@ -2,7 +2,7 @@ import { MSG } from '@/constants/messages';
 import { TABLE } from '@/constants/supabase-tables-name';
 import { fetchUserSessionState } from '@/lib/utils/api/auth-client.api';
 import { supabase } from '@/lib/utils/supabase/supabase-client';
-import type { GetMyGonggamPostsResponse, GonggamPostWithCounts, SortBy } from '@/types/gonggam-posts';
+import type { GetMyGonggamPostsResponse, GonggamPostWithCounts, SortBy } from '@/types/gonggam';
 
 /**
  * Supabase에서 `gonggam_posts_with_counts` 뷰 전체 조회하는 함수
