--- src/types/components/auth-form.ts
@@ -1,5 +1,0 @@
-export interface AuthFormButtonProps {
-  isValid: boolean;
-  isLoginPending?: boolean;
-  isSignupPending?: boolean;
-}
