--- src/types/components/edit-profile-form.ts
@@ -1,11 +1,0 @@
-import type { Tables } from '@/types/supabase';
-
-export interface EditFormData {
-  nickname: string;
-  profile_image_file: FileList | null;
-  profile_image: string;
-}
-
-export interface InitProfile {
-  initProfile: Tables<'users'>;
-}
