+++ src/types/main-banner.ts
