+++ src/types/my-missions.ts
@@ -1,14 +1,4 @@
-import { Database } from '@/types/supabase';
-
-/**
- * 우리가 사용하는 5개의 미션 카테고리만 명시적으로 타입화
- */
-export type UsedTags = '혼밥' | '혼자여행' | '갓생' | '혼놀' | '청소';
-
-/**
- * Supabase에서 mission_list 테이블의 Row 전체 타입
- */
-export type RawMissionsFromSupabase = Database['public']['Tables']['mission_list']['Row'];
+import type { EnumChecklist } from '@/types/supabase-const';
 
 /**
  * Supabase의 user_mission 테이블에서 join된 mission_list 컬럼 구조만 따로 정의
@@ -27,6 +17,6 @@ export type UserMissionRow = {
  * - level은 string으로 유지 (계산 시 parseInt로 처리)
  */
 export type Mission = {
-  type: UsedTags;
+  type: EnumChecklist;
   level: string;
 };
